requirements.txt
# Core
httpx[http2]>=0.27.0
sqlalchemy>=2.0.0
alembic>=1.13.0
apscheduler>=3.10.0
pytz>=2024.1

# Web API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Configuration
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Logging & Utils
loguru>=0.7.0
tenacity>=8.2.0

# Optional CLI
click>=8.1.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
ruff>=0.1.0
mypy>=1.5.0
